qa: ensure tests run correctly when run by root
root is always able to write to a file, UNLESS it is marked as immutable.
As such, I've altered the `testFileNotWritable` test to use `chattr +i` on the test asset file when the root user is detected, so that the root user cannot write to it.
tearDown now also uses `chattr -i` before deleting the test asset file if the root user is detected.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>